example_project/requirements.txt
django-crispy-forms==1.7.2